refactor: OrderSummaryTableViewCell 추가 후 로직 분리
1. OrderSummaryTableViewCell 파일 생성
2. StackTableViewCell 로직 분리
